To run this run the UserName under the WordGuess\src\default then open the UserName.java file.
The way this runs is that the UserName.java asks you to input a user
name. then when one is put in it calls the NewGameOrJoinGame.java under
the GUI package. This asks if you want to run a new game or join a game.
These options aren't fully working obviously because the game doesn't
have that ability to yet. This class will call the MainFrame.java in the
GUI package which will call the rest of the classes for the game play.

~Chris Fickess~